🌌🌍 <3ThomasBS_JegElskerDig!! [ Windows Desktop ] netnode Web Server 🌌🌍 <3ThomasBS_JegElskerDig!!
<3tbs, Mummi and bilbo!!

Thomas Borregaard Sørensen \infinity,-0.16091989,\infinity ONE-MAN ABSOLUTE <3!! I love you, by ???-0.02041977-???write my history please make me please create me for you for me for you for me Camilo Sasuke Thomas Borregaard Sørensen!!

Thomas 3 private commits on mid Dec2020!!

Thomas Online YouTube VODs contribution!!

Mummi right after arm surgery March12!!

Mummi bearing me... :(!!

Mummi (tinytaura) watching and chatting contribution!!

Mummi (tinytaura) at cloud.camilothomas.com machine and at ca2software@twitch channel contribution!!

bilbo sleeping and needing/requesting/crying for help care (for the right person (me), the cats wanna fight with him) contribution!!

sodapoppin and friends contribution!!

iAssyrian chatting contribution!!

alstroemeria5 talk about teaching coding contribution!!

spoofhx (boflux, Spoofh, Benjamin Kuhl) chatting contribution!!

Tectuma (Goth, Tec, Chronos and Promazeus) support, chatting and friendship contribution!!

therealchriff chatting contribution!!

Spoofhx, Gothiccat1 and therealchriff chatting contribution!! 2021-07-19 23:00 BRT

jusg_fpga (fpga_guru, vue_equalizer, just_fpga, Oliver Pohl) chatting contribution!!

TheVibeCurator (Zach.@Discord) chatting contribution!!

cmgriffing streaming contribution!!

smokesess chatting contribution!!

zakariawaseem chatting contribution!!

Pocketwatch (Joseph Riedel 'nd Andy Schatz) streaming contribution!!

SoWeq streaming contribution!!

Hunterpug chatting contribution!!

TimBeaudet (Friends: FletcherLabs, tsjost and Jabokoe) streaming contribution!!

Stumpen_nicklas_dk, sodapoppin and EduardoRFS streaming contribution!!

Roxkstar74 sleeping streaming contribution!!

Promic2 streaming & Retro & Chat games contribution!!

kissloryshy chatting contribution!!

PaperArmys chatting contribution and interest-on-topic:Local_home_web_server contribution!!

blackjekko from Padova Italia through twitch C++/ca2 interest contribution!!

j_blow streaming contribution!!

lowluds chatting and channel interest contribution!!

boflux (Ben, Spoofh, from Germany) chatting contribution!!

parrot_rl chatting contribution (from New Jersey)!!

JPCdk streaming contribution!!

fast_goal chatting and programming interest contribution!!

armtek9 chatting and channel interest contribution!!

ravi387, 0xkira07 and Daytuz chatting contribution!!

BrianFromDenmark subathon streaming contribution!!

Sluggero streaming contribution!!

philliptrudeau streaming, raiding & raiders contribution!!

kampfmaschine_ chatting contribution!!

TehBeardedGamer streaming contribution!!

whyyyyyyysoserious streaming chess contribution!!

fpga_guru (vue_equalizer, Oliver from Deutsch) C++/ca2 interest contribution!!

megadamage streaming contribution!!

camandcode streaming, chatting and production contribution!!

sanskee3 (Cassio@Guarulhos) C++ studying for ca2 development and chatting, testing, reporting and interacting contribution!!

belago68 (@Spain) chatting contributin!!

pqpvoltei ("Guilherme") chatting contribution!!

SovereignDev with Unreal streaming contribution!!

TimBeaudet C++ development and streaming contribution!!

Ash_F0x streaming contribution!!

Myrkee (Valheim) streaming contribution!!

xmetrix and EinfachUwe42 streaming contribution!!

JessicaMak and marcobrunodev streaming contribution!!

alfredotigolo, mandrakenk and Okbatgames chatting contribution!!

jitspoe and Endesga streaming contribution!!

Fearitself (& team mates) streaming contribution!!

BoppyGames dedicated long streaming contribution!!

EZ_username chatting contribution!! (Screen Reader at Linux : AT-SPI2, at Windows MSAA/IA2)

CrashLima (BR) chatting contribution!!

jmcmorris (Jason Morris, SiegeGames) streaming contribution!!

tomrandall (randy) streaming Ludum contribution!!

anas084 (from Saudi Arabia) chatting contribution!!

DonHo streaming contribution!!

vue_equalizer (fpga_guru) chatting contribution!!

Thiagovgamg chatting contribution!!

Naysayer88 and friends contribution!!

lelandkwong streaming contribution!!

Goldbargames streaming contribution!!

Bytakos (bytakos) streaming contribution!!

strager streaming contribution!!

ryanfleury streaming C programming contribution!!

Endesga streaming contribution!!

jitspoe streaming contribution!!

Ash_F0x and JessicaMak streaming contribution!!

WTSRetro/SpiffyDane and Myrkee streaming contribution!!

Ninja and friends streaming contribution!!

erald_guri chatting contribution!!

lastmiles streaming farwest contribution!!

CronicSick (& family) streaming contribution!!

rw_grim streaming contribution!!

rwrxrob broadcast contribution!!

MarckozHD streaming contribution (in 480p though)!!

AdamCYounis streaming contribution!!

unicorndev (and Friends: metrix and Zultralord) streaming and chatting contribution!!

Dunno (P4ndaExpress) chatting and streaming contribution!!

Zorchenhimer streaming contribution!!

lasteveq4 C++ interest chat contriubtion!!

cecilphillip and clarkio @"Microsoft Developer" streaming contribution!!

oijtx streaming contribution!!

Bogdanoff_Chad (diegobrando_linux, Bl4ck_gookoo) chatting (and C++/ca2 interest) contribution!!

intelijens chatting contribution (after long time!!)!!

erald_guri123 chatting contribution (C GUI from scratch at Windows contribution)!!

jhovgaard streaming contribution!!

Klay4_ chatting contribution!!

LinusTech (& luke_afra) streaming contribution!!

HonestDanGames streaming contribution!!

NorthSeaHero streaming contribution!!

Trainwreckstv and friends streaming contribution!!

GoPirateSoftware streaming contribution!!

taiyoinoue, RetroMMO, OfficialAndyPyro and david_joffe streaming contribution!!

Tjienta streaming contribution!!

ThePrimeagen streaming contribution!!

mike_the_swedish_coder (from Sweden) chatting contribution!!

ghostfasm (epryka, liefasm, Vladimir - Vova from Russia) streaming contribuion!!

Jaxstyle and friends streaming contribution!!

EduardRFS streaming contribution!!

togglebit, GexYT streaming contribution!!

phanxgames (Henri Price) streaming contribution!!

qpHalcy0n (and Friends: DragonSlayer0531) streaming and Linux and C++ development contribution!!

Melchizedek6809 and btcfly streaming contribution!!

Llama0x0 and sov_l chatting contribution!!

LLamabit chatting contribution!!

heroseh gamedev (and C/C++ text editor) streaming contribution!!

TaleLearnCode streaming contribution!!

hvalen_hvalborg112 streaming contribution!!

Caro and lus contribution!!

Kebabskal streaming contribution!!

J0KERZZ IRL truck streaming contribution!!

Cristofer Tonita ca2 house wood & general renovation contribution!!

Jean, Eder & family ca2 house renovation & painting contribution!!

zeeth (&friends) streaming contribution!!

tyuuutv chatting contribution (non programmer talking programming related stuff casually)!!

klintcsgo (CS GO: Counter-Strike Global Offensive) streaming contribution!!

løtchoh friendship contribution!!

VerFX chatting contribution!!

KASPERPURE (Super Mario 64) streaming contribution!!

SomewhatAccurate C++ streaming contribution!!

Crinkem (from Australia) chatting, friendship and support contribution!!

chr1sjw (C++, CCTV) streaming contribution!!

TurtleWolfe chatting, hosting and raid contribution!!

ultristors chatting contribution!!

Holmti (hvalen_hvalborg112 friend) streaming contribution!!

Moulberry2 streaming contribution!!

Listening to Bryan Adams, Westlife, Shayne Ward, MLTR, Backstreet Boys, Boyzone - Best Love Songs Ever by Relax Song at YouTube!!

-- hi5 contribution...!!

at Windows 10 Pro at eddie machine remotely from bilbo machine running Windows 10 Pro!!

+++ windowing_win32/mouse_hook.cpp
@@ -127,7 +127,7 @@ namespace mouse_hook
 
       g_pmatter = nullptr;
 
-      PostThreadMessage(g_itask, WM_QUIT, 0, 0);
+      PostThreadMessage((DWORD) g_itask, WM_QUIT, 0, 0);
 
       g_bRun = false;
